requirements.txt

# Core ML Dependencies
torch>=1.13.0
torchvision>=0.14.0
opencv-python>=4.7.0
pillow>=10.1.0
numpy>=1.21.0

# Google Cloud Dependencies
google-cloud-aiplatform>=1.35.0
google-cloud-storage>=2.10.0
google-auth>=2.20.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Optional: Development and testing
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0